fix: NORM
